Refactor navigation into dedicated flow navigators
